Adding credit card processing to the member form.
This is using the stripe.com API. An official account needs to be
created for this application. None of the API key information needs to
be shared with anyone but the sole owner of the application. Once the
account is created, you can use the keys to activate the stripe code.
While in development, all cards will pass.

To use the API key information, start your server with the following:

  STRIPE_API_KEY=iridetheshortbus ruby script/rails s

To use the key on a Heroku production site:

  $ heroku config:add STRIPE_API_KEY=iridetheshortbus
  $ heroku config:add SSL_HOST=usrshare.herokuapp.com